Extract telemetry types into core/types/Telemetry.ts
Move the telemetry/observability type vocabulary out of the catch-all
Interfaces.ts (and MetricsService.ts) into a dedicated core/types/Telemetry.ts,
grouped by subsystem: call context (CallContextLike, InitContext), tracing
(SpanKind, SpanConfig, FullSpanConfig), metrics (MetricTags), and activity
tracking (TrackSeverity, TrackOptions). Re-exported via the core barrel.

No behavior change. Public import path is unchanged for @xh/hoist/core
consumers; MetricTags moves from the @xh/hoist/svc export to @xh/hoist/core
(new on this branch, only consumer is MetricsService).
